+++ scripts/loop/extract_stoch_loop_TSM_h5.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-if [ $# -ne 10 ]
+if [ $# -ne 11 ]
 then
-    echo "Usage: $0 <h5-file> <extract_dir> <out_dir> <mom_list> <stoch_list> <LowPrecSum:1, HighPrecSum:0> <HighPrec, LowPrec> <exe_dir> <T> <traj>"
+    echo "Usage: $0 <h5-file> <extract_dir> <out_dir> <mom_list> <stoch_list> <LowPrecSum:1, HighPrecSum:0> <HighPrec, LowPrec> <exe_dir> <T> <traj> <HighMomForm:1 LowMomForm:0>"
     exit
 fi
 
@@ -17,41 +17,71 @@ PREC_TYPE=$7
 EXE_DIR=$8
 T=$9
 TRAJ=${10}
+HighMomForm=${11}
 
 EX_DIR=${DIR}/stoch_hdf5_extracted
 
-mkdir -p ${EX_DIR}
 mkdir -p ${OUT_DIR}
 
-if [ ${SUM_TYPE} -eq 1 ]
+if [ $HighMomForm -eq 0 ]
 then
-    STRING=NLP
-fi
-if [ ${SUM_TYPE} -eq 0 ]
-then
-    STRING=${PREC_TYPE}_NHP
-fi
 
+    mkdir -p ${EX_DIR}
 
-for tp in Scalar dOp Loops LoopsCv LpsDw LpsDwCv
-do
-    for Ns in `cat $NSTOCH_LIST`
+    if [ ${SUM_TYPE} -eq 1 ]
+    then
+	STRING=NLP
+    fi
+    if [ ${SUM_TYPE} -eq 0 ]
+    then
+	STRING=${PREC_TYPE}_NHP
+    fi
+
+    for tp in Scalar dOp Loops LoopsCv LpsDw LpsDwCv
     do
-        OUT_FILE=${OUT_DIR}/loop_stoch.${TRAJ}_TSM_${STRING}${Ns}_${tp}.loop.dat
-	rm -f ${OUT_FILE}
-
-	while read mx my mz
+	for Ns in `cat $NSTOCH_LIST`
 	do
-	    EX_FILE=${EX_DIR}/loop_stoch.${TRAJ}_${STRING}${Ns}_${tp}.${mx}_${my}_${mz}.loop.dat
-	                      
-	    C="${EXE_DIR}/extract_stoch_loop_TSM_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${TRAJ} ${Ns} ${SUM_TYPE} ${T}"
-#	    echo $C
+            OUT_FILE=${OUT_DIR}/loop_stoch.${TRAJ}_TSM_${STRING}${Ns}_${tp}.loop.dat
+	    rm -f ${OUT_FILE}
 
-	    eval $C
+	    while read mx my mz
+	    do
+		EX_FILE=${EX_DIR}/loop_stoch.${TRAJ}_${STRING}${Ns}_${tp}.${mx}_${my}_${mz}.loop.dat
+	        
+		${EXE_DIR}/extract_stoch_loop_TSM_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${TRAJ} ${Ns} ${SUM_TYPE} ${T}
 
-	    cat ${EX_FILE} >> ${OUT_FILE}
-	done < ${MOM_LIST}
+		cat ${EX_FILE} >> ${OUT_FILE}
+	    done < ${MOM_LIST}
+	done
     done
-done
 
-#rm -rf ${EX_DIR}
+    rm -rf ${EX_DIR}
+
+fi
+
+
+if [ $HighMomForm -eq 1 ]
+then
+
+    if [ ${SUM_TYPE} -eq 1 ]
+    then
+	STRING=NLP
+    fi
+    if [ ${SUM_TYPE} -eq 0 ]
+    then
+	STRING=${PREC_TYPE}_NHP
+    fi
+
+    for tp in Scalar dOp Loops LoopsCv LpsDw LpsDwCv
+    do
+	for Ns in `cat $NSTOCH_LIST`
+	do
+            OUT_FILE=${OUT_DIR}/loop_stoch.${TRAJ}_TSM_${STRING}${Ns}_${tp}.loop.dat
+	    rm -f ${OUT_FILE}
+	        
+	    ${EXE_DIR}/extract_stoch_loop_TSM_HighMomForm_h5 ${FILE} ${OUT_FILE} ${tp} ${TRAJ} ${Ns} ${SUM_TYPE} ${T}
+		
+	done
+    done
+
+fi
